Extract v1/v2 key forming conventions to a mixin
